+++ CHANGELOG.md
@@ -8,3 +8,9 @@
 =======
 
 - added a quasiquoter in a new public library
+
+1.1.0.0
+=======
+
+- removed deprecated functions fieldAliases and branchAliases
+- renamed ForRubric to AliasType